Add recipe for creating the docker container the test will run.
We need a container that has the following properties:

 - Some sort of verifiable output for `docker logs`
 - Runs until `docker stop`

We do this with a simple shell loop that has defined content at
the beginning and end that can be used for verification, and dynamic
content (the timestamp every second) while the container is running to
give humans a frame of reference.

We choose alpine as the base because it is light weight.