Add a tar of the wellknown image to the python distribution.
In order to accomplish my goal of "test code doesn't need to hit the
network" I need a way for the test to get a copy of the container that
should be tested.  Thankfully docker can import from a tar!

Thus as part of the build we generate wellknown.tar, and add that to the
python packaging manifest.

Since this is an intermediate build artifact, it gets .gitignored.