Add auto json conversion QoL features to our client.
With this pass, if a request is passed a dict as the body, it will
automatically serialize that to json.  This saves a bit of pain for
consumers.  Similarly, response objects now have a .json() to
automatically convert a json body to its Python datastructure
representation using json.parse.

Not strictly necessary, but these make the client more pleasant to use.