Add a demux_logs helper function.
Docker's logs handling has stdout and stderr interlaced as a stream of
8 byte headers followed by body (repeat header-body until stream end).

This patch adds a function to demux these raw bytestreams into two
strings for stdout and stderr, which is what our test code will want to
interact with.

I've added some limited tests for demux_logs as that was easier than
trying to develop it on live data.