Bind raw test functions to unittest.TestCase.
Instead of executing the test lifecycle as a simple script, I now use
the unittest harness to have more robust logging and error handling
(e.g. teardown will be run even if the test has an error).

I've also added a main file and function, with the intent of having that
be the entrypoint (as defined in setup.py) to this code.

Addresses:
    https://github.com/javins/gravitational-audition/issues/8
and
    https://github.com/javins/gravitational-audition/pull/5#discussion_r335759224