Add .body property to FriendlyHttpResponse.
In order to do logging well, we want to log response bodies.  However,
if a test has already called .read() on the response, the body is not
available to logging without passing it.

This is a bit of a hack to cache the body in the response object, so
that it can be accessed multiple times.  I believe the requests
response object does something like this.