Validate stdout after stderr.
If stderr is non-empty, we might expect the stdout stream to be
corrupted some how, as there has been an error.  Thus check stderr
first, so that it will show up in the test logs, and potentially be
useful to debugging.

Addresses https://github.com/javins/gravitational-audition/pull/9#discussion_r337121072