Add dev-tools and install targets.
In order to make buidling and running this project as easy as possible,
add `make install` to install, and the `make dev-tools` target to
install all the tools necessary to develop this project.

This commit also formalizes which version of tools are required in
dev-requirements.txt, to make sure that all developers have a common
toolchain.