Add sudo to readme commands.
Without this the pip installs `grav-api-test` to `~/.local/bin`, and I'd
just as soon not have to try to explan that in the readme.

The proper thing to do here is use a .deb as our distribution format,
but that is a signifigant amount of new code I'd just as soon not write.